update API spec for token endpoints
