+++ components/renku_data_services/connected_services/apispec.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2026-01-26T13:18:55+00:00
+#   timestamp: 2026-02-05T09:28:04+00:00
 
 from __future__ import annotations
 
@@ -10,6 +10,19 @@ from typing import List, Optional
 
 from pydantic import ConfigDict, Field, RootModel
 from renku_data_services.connected_services.apispec_base import BaseAPISpec
+
+
+class OAuth2Token(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="allow",
+    )
+    access_token: Optional[str] = Field(
+        None, description="An access token for OAuth 2.0"
+    )
+    scope: Optional[str] = None
+    token_type: Optional[str] = None
+    id_token: Optional[str] = None
+    expires_at_iso: Optional[str] = None
 
 
 class RepositorySelection(Enum):
@@ -26,6 +39,18 @@ class AppInstallation(BaseAPISpec):
     account_web_url: str
     repository_selection: RepositorySelection
     suspended_at: Optional[datetime] = None
+
+
+class PostTokenResponse(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="allow",
+    )
+    access_token: str
+    token_type: str
+    expires_in: int
+    refresh_token: str
+    refresh_expires_in: Optional[int] = None
+    scope: Optional[str] = None
 
 
 class ProviderKind(Enum):
@@ -49,6 +74,10 @@ class PaginationRequest(BaseAPISpec):
     per_page: int = Field(
         20, description="The number of results per page", ge=1, le=100
     )
+
+
+class PostTokenGrantType(Enum):
+    refresh_token = "refresh_token"
 
 
 class Error(BaseAPISpec):
@@ -249,8 +278,8 @@ class ConnectedAccount(BaseAPISpec):
         extra="forbid",
     )
     username: str = Field(..., examples=["some-username"])
-    web_url: str = Field(
-        ...,
+    web_url: Optional[str] = Field(
+        None,
         description="A URL which can be opened in a browser, i.e. a web page.",
         examples=["https://example.org"],
     )
@@ -258,6 +287,14 @@ class ConnectedAccount(BaseAPISpec):
 
 class AppInstallationList(RootModel[List[AppInstallation]]):
     root: List[AppInstallation]
+
+
+class PostTokenRequest(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="allow",
+    )
+    grant_type: PostTokenGrantType
+    refresh_token: str
 
 
 class ProviderList(RootModel[List[Provider]]):
